Arreglos a los modelos y quickfix al contrato de la api

+++ app/src/main/java/com/wolfteam20/schedulemobile/data/network/models/ProfesorDTO.kt
@@ -13,5 +13,5 @@ data class ProfesorDTO(
     var cedula: Long,
     var nombre: String,
     var apellido: String,
-    var idPrioridad: Int
+    var idPrioridad: Long
 ):Parcelable
